move deploymenttemplateservice.go to pkg/deployment/manifest/deploymentTemplate as it makes sense for all depl template service layer methods to be at one place , this is part one of such refactoring
