+++ pkg/deployment/manifest/deploymentTemplate/wire_deploymentTemplate.go
@@ -18,6 +18,7 @@ package deploymentTemplate
 
 import (
 	"github.com/devtron-labs/devtron/pkg/deployment/manifest/deploymentTemplate/chartRef"
+	"github.com/devtron-labs/devtron/pkg/deployment/manifest/deploymentTemplate/read"
 	"github.com/google/wire"
 )
 
@@ -28,4 +29,6 @@ var DeploymentTemplateWireSet = wire.NewSet(
 	wire.Bind(new(DeploymentTemplateValidationService), new(*DeploymentTemplateValidationServiceImpl)),
 	chartRef.NewChartRefServiceImpl,
 	wire.Bind(new(chartRef.ChartRefService), new(*chartRef.ChartRefServiceImpl)),
+	read.NewDeploymentTemplateHistoryReadServiceImpl,
+	wire.Bind(new(read.DeploymentTemplateHistoryReadService), new(*read.DeploymentTemplateHistoryReadServiceImpl)),
 )
